web_client/requirements.txt
flask
requests
psycopg2-binary
sqlalchemy
flask-sqlalchemy
flask-login
python-dotenv
